// generated/src/app/api-models/cap-control-dtl.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapControlDtl {
  capModel: string;
  firstContractMonth: Date;
  lastContractMonth: Date;
  runDate: Date;
  capMonthRun: Date;
  openClosed: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}